requirements.txt
pandas==1.5.0
joblib
streamlit
numpy
scikit-learn
openpyxl
xlsxwriter
seaborn
python-binance
mplfinance